[APC-6279] Resize the rootfs volume before copying 2nd partition

Closes APC-6279

See merge request public-projects/yocto/meta-iris-base!240

+++ recipes-support/swupdate/swupdate/power_on_selftest.sh
@@ -63,6 +63,20 @@ update_alternative_firmware() {
 	rm -r /tmp/cur_fitimage_dev /tmp/alt_fitimage_dev
 	[ "$err" -eq 0 ] || return "$err"
 
+	# Resize the rootfs volume before copying
+	# Read new volume size from current rootfs volume (trim leading spaces from lvs output), compare with size of alternative rootfs volume
+	new_size=$(lvs "/dev/mapper/irma6lvm-rootfs_${CUR_FW_SUFFIX}" -o LV_SIZE --noheadings --units B --nosuffix | sed 's/^ *//g')
+	alt_size=$(lvs "/dev/mapper/irma6lvm-rootfs_${ALT_FW_SUFFIX}" -o LV_SIZE --noheadings --units B --nosuffix | sed 's/^ *//g')
+
+	# Resize alternate rootfs volume
+	if [ "$new_size" != "$alt_size" ]; then
+		lvresize --autobackup n --force --yes --quiet -L "$new_size"B "/dev/mapper/irma6lvm-rootfs_${ALT_FW_SUFFIX}" 2> /dev/null || \
+			{ log "Error: Failed to resize alternative rootfs volume"; err=1; }
+	fi
+	vgmknodes
+
+	[ "$err" -eq 0 ] || return "$err"
+
 	# Update alternative rootfs and rootfs hashtree volumes
 	dd if="/dev/mapper/irma6lvm-rootfs_${CUR_FW_SUFFIX}" of="/dev/mapper/irma6lvm-rootfs_${ALT_FW_SUFFIX}" bs=10M >/dev/null 2>&1 || \
 		{ log "Error: Failed to copy alternative rootfs"; err=1; return "$err"; }
@@ -84,6 +98,7 @@ prepare_alternative_fw_update() {
 	mount /dev/mapper/irma6lvm-keystore /mnt/keystore
 	if ! cmp -s /mnt/keystore/rootfs_a_roothash /mnt/keystore/rootfs_b_roothash; then
 		touch "$ALTERNATIVE_FW_UPDATE_FLAG"
+		sync
 	fi
 	umount /mnt/keystore
 }
@@ -101,14 +116,13 @@ reset_uboot_envs() {
 PENDING_UPDATE=$(fw_printenv upgrade_available | awk -F'=' '{print $2}')
 if [ "$PENDING_UPDATE" = "1" ]; then
 	power_on_selftest
-	prepare_alternative_fw_update
 	reset_uboot_envs
+	prepare_alternative_fw_update
 	log "Firmware update successful complete"
 fi
 
 # Update the alternative firmware after success
 if [ -f "$ALTERNATIVE_FW_UPDATE_FLAG" ]; then
-	update_alternative_firmware && log "Alternative firmware update complete" || log "Alternative firmware update failed"
-	rm "$ALTERNATIVE_FW_UPDATE_FLAG"
+	update_alternative_firmware && rm "$ALTERNATIVE_FW_UPDATE_FLAG" && log "Alternative firmware update complete" || log "Alternative firmware update failed"
 fi
 } &
